api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai==1.82.0
mcp==1.9.2
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.3
